CancelIoEx cancels the file, not merely the handle.
Well, it is ambiguous either way.
It cancels for the underlying file object.
Two CreateFile,s is two objects.
DuplicateHandle, is not. I think.
Notice that CancelIo w/o Ex is similar, and says "file", not "file handle".
Nothing is clear here, granted. But I looked at the code to determine if it is was file object or handle.
Handle would have been surprising.